move notify functions to notify.rs
